pyproject.toml
[project]
name = "mailbudgettracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Matthew Wisdom",email = "matthewwisdom11@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "apiclient (>=1.0.4,<2.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-api-python-client (>=2.160.0,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "coverage (>=7.6.10,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.9.4"
